requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cymem==2.0.11
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
keras==3.11.3
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
murmurhash==1.0.13
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
preshed==3.0.10
protobuf==4.25.8
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.16.0
thinc==8.2.5
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3